Custom parameters

+++ portmanager-backend/src/main/java/com/portmanager/repository/TerminalRepository.java
@@ -3,4 +3,4 @@ package com.portmanager.repository;
 import com.portmanager.entity.TerminalEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TerminalRepository extends JpaRepository<TerminalEntity, Integer> {}
+public interface TerminalRepository extends JpaRepository<TerminalEntity, Long> {}
